requirements.txt
flask==3.1.0
pandas==2.2.3
openpyxl==3.1.5
beautifulsoup4==4.13.3
requests==2.32.3
lxml==5.3.1
html5lib==1.1
googlesearch-python==1.3.0
aiohttp==3.11.12
werkzeug==3.1.3
gunicorn==22.0.0
